+++ release_notes.md
@@ -4,6 +4,8 @@
 
 * Add `loglevel` setting to [app] section (default is Warning)
 * Add `validate` interactive command to check `uptime.conf`
+* Improve tag splitting (opened related [bug report](https://github.com/dotnet/diagnostics/issues/5935))
+* Silently drop dupe tags like `pid`; log Warning if mismatched
 * Show Windows permissions reminder when running service as user
 * README warning about provider-wildcard overhead and filtering
 * Bugfix - provider process-filter